pyproject.toml
[project]
name = "agent"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0",
    "langchain>=1.0",
    "langchain-openai>=1.0",
    "langgraph>=1.0",
    "langgraph-checkpoint-sqlite>=2.0",
    "langgraph-cli[inmem]>=0.3",
    "supabase>=2.9",
    "twilio>=9.3",
    "pydantic>=2.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]
